Extract jsx from main logic in table rows
